+++ CppSource/SHT35/sht35.cpp
@@ -9,11 +9,9 @@ uint16_t SHT35::readTemperature() {
 	std::uniform_real_distribution<double> noise(-0.25, 0.25); // Adjust the range as needed
 
 	// Add noise to the temperature within the specified deviation
-	double noisyTemperature = greenhouse.getTemperature() + noise(gen);
-
+	float noisyTemperature = greenhouse.getTemperature() + float(noise(gen));
 	uint16_t rawData = 0x0000;
-
-	rawData = int((noisyTemperature + 45) * (65535 / 175));
+	rawData = uint16_t((noisyTemperature + 45) * (65535 / 175));
 
 	return rawData;
 }
@@ -24,11 +22,10 @@ uint16_t SHT35::readHumidity() {
     std::uniform_real_distribution<double> noise(-1.5, 1.5); // Adjust the range as needed
 
     // Add noise to the humidity within the specified deviation
-    double noisyHumidity = greenhouse.getHumidity() + noise(gen);
-
+    float noisyHumidity = greenhouse.getHumidity() + float(noise(gen));
     uint16_t rawData = 0x0000;
 
-    rawData = int((65535 / 100) * noisyHumidity);
+    rawData = uint16_t((65535 / 100) * noisyHumidity);
     return rawData;
 }
